+++ src/layouts/InterfaceLayout/components/InterfaceTokens/InterfaceTokens.vue
@@ -72,7 +72,12 @@
                   "
                   rel="noopener noreferrer"
                   target="_blank"
-                  >{{ token.symbol }}</a
+                >
+                  <span>{{
+                    token.symbol.length > 10
+                      ? `${token.symbol.substr(0, 15)}...`
+                      : token.symbol
+                  }}</span></a
                 >
               </td>
               <td
@@ -83,7 +88,11 @@
                 {{ token.balance }}
               </td>
               <td v-else>
-                {{ token.balance }}
+                {{
+                  token.balance.length > 10
+                    ? `${token.balance.substr(0, 25)}...`
+                    : token.balance
+                }}
                 <i
                   class="fa fa-refresh clickable"
                   @click="getSpecificTokenBalance(token)"