+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### Release v5.7.23-hotfix.1
+
+### Bug
+
+* Fix weird token issue when tokens has way too long of a symbol [#2894](https://github.com/MyEtherWallet/MyEtherWallet/pull/2894)
 ### Release v5.7.23
 
 ### Bug